+++ libs/render_delegate/render_pass.cpp
@@ -298,7 +298,7 @@ AtNode* _CreateFilter(HdArnoldRenderDelegate* renderDelegate, const HdAovSetting
     // We need to make sure that it's holding a string, then try to create it to make sure
     // it's a node type supported by Arnold.
     const auto filterType = _GetOptionalSetting(aovSettings, _tokens->aovSettingFilter, std::string{});
-    if (filterType.empty() || filterType == "cryptomatte_filter") {
+    if (filterType.empty()) {
         return nullptr;
     }
     const auto filterNameStr =
@@ -530,6 +530,7 @@ void HdArnoldRenderPass::_Execute(const HdRenderPassStateSharedPtr& renderPassSt
         }
     }
     GfVec4f windowNDC = _renderDelegate->GetWindowNDC();
+    float pixelAspectRatio = _renderDelegate->GetPixelAspectRatio();
     // check if we have a non-default window
     bool hasWindowNDC = (!GfIsClose(windowNDC[0], 0.0f, AI_EPSILON)) || 
                         (!GfIsClose(windowNDC[1], 0.0f, AI_EPSILON)) || 
@@ -566,7 +567,6 @@ void HdArnoldRenderPass::_Execute(const HdRenderPassStateSharedPtr& renderPassSt
 
     if (windowChanged) {
         renderParam->Interrupt(true, false);
-        AiNodeResetParameter(options, str::pixel_aspect_ratio);
         if (hasWindowNDC) {
             _windowNDC = windowNDC;
             
@@ -626,8 +626,7 @@ void HdArnoldRenderPass::_Execute(const HdRenderPassStateSharedPtr& renderPassSt
 
                 // For interactive renders, need to adjust the pixel aspect ratio to match the window NDC
                 if (!_renderDelegate->IsBatchContext()) {
-                    float pixel_aspect_ratio = xDelta / yDelta;
-                    AiNodeSetFlt(options, str::pixel_aspect_ratio, pixel_aspect_ratio);
+                    pixelAspectRatio *= xDelta / yDelta;
                 }
             
             } else {
@@ -650,6 +649,7 @@ void HdArnoldRenderPass::_Execute(const HdRenderPassStateSharedPtr& renderPassSt
             _windowNDC = GfVec4f(0.f, 0.f, 1.f, 1.f);
         }
     }
+    AiNodeSetFlt(options, str::pixel_aspect_ratio, pixelAspectRatio);
 
     auto checkShader = [&] (AtNode* shader, const AtString& paramName) {
         auto* options = _renderDelegate->GetOptions();
@@ -805,7 +805,7 @@ void HdArnoldRenderPass::_Execute(const HdRenderPassStateSharedPtr& renderPassSt
                 // But Arnold won't recognize this as being the actual beauty and adaptive sampling
                 // won't apply properly (see #1006). So we want to detect which output is the actual beauty 
                 // and treat it as Arnold would expect.
-                bool isBeauty = (sourceName == "C.*") && (binding.aovName == HdAovTokens->color);
+                bool isBeauty = binding.aovName == HdAovTokens->color;
                 
                 // When using a raw buffer, we have special behavior for color, depth and ID. Otherwise we are creating
                 // an aov with the same name. We can't just check for the source name; for example: using a primvar